requirements.txt
# Core
litellm>=1.30.0
openai>=1.12.0
anthropic>=0.25.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn>=0.29.0

# SWE-bench
swebench>=1.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# ML / Embedding
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0

# Config
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.2.0

# Logging & Monitoring
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
